code smells removed

+++ src/main/java/secondmilestone/DataFilter.java
@@ -109,7 +109,7 @@ public class DataFilter {
         }
         
         // percentage of the majority class
-        double percentage = (100 * 2 * numberOfBuggy) / dataset.size();
+        int percentage = (100 * 2 * numberOfBuggy) / dataset.size();
         if (percentage >= 50)
             return percentage;
         else
